M2-6463: Rename LaunchDarkly to FeatureFlags
