+++ src/shared/utils/featureFlags/featureFlags.ts
@@ -1,11 +1,11 @@
 import { LDClient } from 'launchdarkly-react-client-sdk';
 
-import { PROHIBITED_PII_KEYS } from './launchDarkly.const';
+import { PROHIBITED_PII_KEYS } from './featureFlags.const';
 
 let _ldClient: LDClient;
 let _userId: string;
 
-export const LaunchDarkly = {
+export const FeatureFlags = {
   init(client: LDClient) {
     _ldClient = client;
   },
